requirements.txt
Adafruit-Blinka==8.46.0
adafruit-circuitpython-amg88xx==1.2.21
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.1
adafruit-circuitpython-register==1.9.18
adafruit-circuitpython-requests==4.1.4
adafruit-circuitpython-typing==1.10.3
Adafruit-PlatformDetect==3.72.0
Adafruit-PureIO==1.1.11
appdirs==1.4.4
binho-host-adapter==0.1.6
blessed==1.20.0
contourpy==1.2.1
cycler==0.12.1
flyingcircus==0.1.4.1
fonttools==4.53.1
kiwisolver==1.4.5
matplotlib==3.9.1
numpy==2.0.0
packaging==24.1
pillow==10.4.0
pyftdi==0.55.4
pyparsing==3.1.2
pyserial==3.5
python-dateutil==2.9.0.post0
pytk==0.0.2.1
pyusb==1.2.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
setuptools-scm==8.1.0
six==1.16.0
smbus2==0.4.3
sysv-ipc==1.1.0
typing_extensions==4.12.2
wcwidth==0.2.13
